- Fixed the project on Windows (MSVC Community Edition 2013) - Some little tweak in the Linux Project (Codeblock) - File selector display accented characters again on Windows
git-svn-id: http://svn.code.sf.net/p/irrrpgbuilder/code@649 528fbfa9-60f6-4f4a-a82b-d133dc65ca33

+++ src/gui/GUIManager.cpp
@@ -398,13 +398,13 @@ void GUIManager::setWindowVisible(GUI_CUSTOM_WINDOW window, bool visible)
             mouseX = App::getInstance()->getDevice()->getCursorControl()->getPosition().X-100;
             mouseY = App::getInstance()->getDevice()->getCursorControl()->getPosition().Y-20;
 
-			if (visible && screen.Height-200<mouseY+20) // Reposition the menu if it will be cropped by the screen clipping
+			if (visible && int(screen.Height-u32(200))<mouseY+20) // Reposition the menu if it will be cropped by the screen clipping
 				mouseY-=160;
 
 			if (visible && mouseX<50) // Reposition the menu if it will be cropped by the screen clipping
 				mouseX+=100;
 
-			if (visible && (screen.Height-200<App::getInstance()->getDevice()->getCursorControl()->getPosition().Y || (App::getInstance()->getDevice()->getCursorControl()->getPosition().X-100)<50))
+			if ((visible && int(screen.Height-u32(200))<App::getInstance()->getDevice()->getCursorControl()->getPosition().Y || (App::getInstance()->getDevice()->getCursorControl()->getPosition().X-100)<50))
 			{
 				device->getCursorControl()->setPosition(mouseX+100,mouseY+20);
 			}
@@ -817,6 +817,7 @@ void GUIManager::loadFonts()
 	guiFont8 = guienv->getFont("../media/fonts/Arial8.xml");
 	guiFont9 = guienv->getFont("../media/fonts/Trebuchet10.xml");
 	guiFont10 = guienv->getFont("../media/fonts/Arial10.xml");
+	//guiFont10 = guienv->getFont("../media/fonts/Trebuchet10.xml");
 	guiFont12 = guienv->getFont("../media/fonts/Arial12.xml");
 	guiFont14 = guienv->getFont("../media/fonts/Arial14.xml");
